pyproject.toml
[tool.poetry]
name = "showcase_3d_printer_anomaly_counter"
version = "0.1.1"
description = "Anomaly Counter per print Request for 3D-Printer Showcase"
license = "Apache-2.0"
authors = ["Joern Richter <joern.richter@th-koeln.de>"]
readme = "README.md"
repository = "https://github.com/innovation-hub-bergisches-rheinland/showcase-3d-printer-anomaly-counter"

[tool.poetry.dependencies]
python = "^3.8"
appdirs = "^1.4.4"
kafka-python = "^2.0.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
bandit = "^1.7.0"
black = "^21.7b0"
flake8 = "^3.9.2"
pylama = "^7.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
